requirements.txt
asgiref==3.5.0
asttokens==2.0.5
autopep8==1.6.0
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.2
debugpy==1.6.0
decorator==5.1.1
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.11.0
django-filter==21.1
django-heroku==0.3.1
django-localflavor==3.1
djangorestframework==3.13.1
drf-yasg==1.20.0
entrypoints==0.4
executing==0.8.3
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
ipykernel==6.13.0
ipython==8.3.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.1
jupyter-client==7.3.0
jupyter-core==4.10.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.29
psutil==5.9.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pure-eval==0.2.2
pycodestyle==2.8.0
Pygments==2.12.0
PyJWT==2.3.0
pyparsing==3.0.8
pypiwin32==223
python-dateutil==2.8.2
python-stdnum==1.17
pytz==2022.1
pywin32==304
pyzmq==22.3.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
tornado==6.1
traitlets==5.1.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
whitenoise==6.0.0
